requirements.txt
py-clob-client>=0.34.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
python-telegram-bot==21.4
pandas>=2.1.0
numpy>=1.26.0
httpx>=0.25.0
